third_party/WebKit/LayoutTests/http/tests/devtools/sources/autocomplete-scss-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
The test verifies autocomplete suggestions for SCSS file.


Running: testPropertyValueSuggestionsBefore$
========= Selection In Editor =========
@mixin my-border-style($style) {
    border-style: |$;
}
======= Autocomplete Suggestions =======
[dashed, dotted, double, groove, hidden, inherit, initial, inset, none, outset, ridge, solid, unset]

Running: testPropertyValueSuggestionsAfter$
========= Selection In Editor =========
@mixin my-border-style($style) {
    border-style: $|;
}
======= Autocomplete Suggestions =======
[$style]
